+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 - Enh: Compiled multiple translation messages into single `base` file
 - Fix #177: CLI error when no REST module is installed
 - Enh #148: Store folding state for categories per user
+- Enh #180: Option to add wiki pages to Space menu
 
 
 1.5.3  (April 16, 2021)